feat(element): add ELEMENT_NOT_FOUND to reducer to remove active elem
if active element is not removed then the app will think the active element was
the prior element when in fact the active is null. Example is loading a new
element type in the designer different from the prior one. If this commit is
not in place it will load the schema for the prior element.